Move corpus test to integration test
