+++ src/lib.rs
@@ -146,80 +146,7 @@ mod test {
     use super::format;
     use miette::{miette, Result};
     use rayon::prelude::*;
-    use std::{
-        collections::HashMap,
-        path::{Path, PathBuf},
-    };
-
-    #[test]
-    fn corpus() -> Result<()> {
-        use pretty_assertions::assert_eq;
-
-        let corpus = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("corpus");
-
-        let update_baseline = std::env::var("UPDATE_BASELINE").is_ok();
-
-        let files = walkdir::WalkDir::new(&corpus)
-            .into_iter()
-            .filter_map(|e| {
-                let e = e.ok()?;
-
-                if e.file_type().is_file()
-                    && e.path().extension().and_then(|ext| ext.to_str()) == Some("spicy")
-                {
-                    Some(e)
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<_>>();
-
-        files
-            .par_iter()
-            .filter_map(|t| {
-                let input = std::fs::read_to_string(t.path()).ok()?;
-
-                let output = {
-                    let path = t.path().to_path_buf();
-
-                    let file_name = format!(
-                        "{}.expected",
-                        path.file_name()
-                            .unwrap_or_else(|| panic!(
-                                "cannot get filename component of {}",
-                                path.display()
-                            ))
-                            .to_string_lossy()
-                    );
-
-                    let mut o = path;
-                    assert!(o.pop());
-
-                    o.join(file_name)
-                };
-
-                let formatted = format(&input, false, false).unwrap_or_else(|_| {
-                    panic!("cannot format source file {}", t.path().to_string_lossy())
-                });
-
-                if !update_baseline {
-                    let expected = std::fs::read_to_string(output).expect("cannot read baseline");
-                    assert_eq!(
-                        expected,
-                        formatted,
-                        "while formatting {}",
-                        t.path().display()
-                    );
-                } else {
-                    std::fs::write(output, formatted).expect("cannot update baseline");
-                }
-
-                Some(1)
-            })
-            .collect::<Vec<_>>();
-
-        Ok(())
-    }
+    use std::{collections::HashMap, path::Path};
 
     #[test]
     fn no_trailing_newline() -> Result<()> {
